backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
pydantic==2.6.1
pydantic-settings==2.2.1
passlib==1.7.4
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
